<aside> <img src="/icons/thought-dialogue_red.svg" alt="/icons/thought-dialogue_red.svg" width="40px" /> **Activity - Componentizing**
- The list of tasks should be its own component.
- Add TasksList.jsx file and create a function component, and move the hard-coded array and the code for rendering the list items from App.jsx to this component.
</aside>
