+++ week9/my-react-app/src/components/Task.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function Task({task}) {
+  return (
+    <li>
+        <div>
+            <p>{task.title}</p>
+            <p>{task.date}</p>
+        </div>
+    </li>
+  );
+}
