+++ week9/my-react-app/src/components/TasksList.jsx
@@ -1,0 +1,31 @@
+import React from 'react'
+import Task from './Task'
+
+export default function TasksList() {
+    const tasks = [
+  {
+    id: 1,
+    title: "Review week 9 material",
+    date: "June 4th at 1 pm",
+  },
+  {
+    id: 2,
+    title: "Do quiz 9",
+    date: "June 4th at 6 pm",
+  },
+  {
+    id: 3,
+    title: "Work on assignment 2",
+    date: "June 5th at 8 am",
+  },
+  ];
+  return (
+    <>
+      <ul>
+        {tasks.map((tasks) => {
+            return <Task key={tasks.id} task={tasks} />;
+        })}
+      </ul>
+    </>
+  )
+}
